Represent DOM elements as data (nested vectors)
Add a lightweight implementation of the "Hiccup"-template language to the project.  This represents HTML elements as nested vectors.

For example, to build a table cell with a combo box, we need the following HTML structure:

    <td>
      <div class="select">
        <select>
          <option>Select</option>
        </select
      </div>
    </td>

Currently, we build using nested JavaScript DOM API calls:

    (dom/createDom "td"
      (dom/createDom "div" {:class  "select"}
        (dom/createDom "select"
          (dom/createDom "option" "Select"))))

Instead of code, construct the same nested structure using data:

    [:td
       [:div.select
         [:select
           [:option "Select"]]]]

This is more concise in most of the cases where multiple levels of nesting is required.

But more importantly, this representation is _just data_, so it can be shared between the frontend and the backend.

In our case we don't have a backend yet, but we pre-render the bulk of the page using the same "template language".