Rewrite the UI components using async processes.
In this PR we continue pursuing the ideal of decoupling Event handling and DOM manipulation from the
core application logic.

To this end we use two fundamental building blocks: lightweight processes (async/go blocks) and
channels.

The general idea is to transform the DOM Events to data and put them on channels. The application
listens on input channels, manipulate the DOM and produce to output channels:

```
DOM Event -> Even processing -> Input Channels -> Async process [DOM Manipulation] -> Output Channels
```

The application logic sits in async processes which access the DOM only via interfaces.  This
provides excellent isolation from the details of the DOM and Events APIs.  The separation is not
always perfect, but it can be easily improved later, if needed, by introducing more redirection via
interfaces.

There are five processes: input, tab, render, table and validator.  Their role is explained in
docstrings.  The most interesting is the table process which processes data from many of events and
it dynamically starts the validator process when the user edits a cell.

One novelty is the Tabs which allows the user to work on the Sufaces, Fields, or on the Aperture,
three different aspects of the optical system.  The Fields and Aperture tabs are currently static,
but with the infrastructure in place they will be easily integrated into the live editing system.