Avoid overindexing when rays terminate
Currently the rays drawing works only if the number of rays is identical on all surfaces.  This was
captured in the `::raytrace-results` spec so it didn't cause any problems, but it also prevented
some results to be drawn.

Check the length of each surface when iterating and remove the `all-same-length` condition from the spec.