Expose the Point Source field type to the frontend
This adds support to the frontend for specifying point source fields.

The biconvex lens example was also added to the frontend.

A major change is that ray directions are no longer required to be unit vectors. Numerical errors in the chain of trigonometric calculations almost inevitably led to ray trace failures for finite object distances.